+++ src/main/java/com/grolinger/java/service/DataProcessorService.java
@@ -15,4 +15,6 @@ public interface DataProcessorService {
     Context processContextOfApplication(String colorName, String integrationType, String applicationName, String serviceName, String interfaceName, Integer orderPrio);
 
     void processApplication(final List<ApplicationDefinition> applicationList, final DiagramType diagramType) throws IOException;
+
+    void exportTemplate();
 }