+++ src/main/java/com/grolinger/java/service/adapter/exportdata/impl/LocalExportAdapterImpl.java
@@ -1,7 +1,7 @@
-package com.grolinger.java.service.adapter.impl;
+package com.grolinger.java.service.adapter.exportdata.impl;
 
 import com.grolinger.java.controller.templatemodel.DiagramType;
-import com.grolinger.java.service.adapter.FileService;
+import com.grolinger.java.service.adapter.exportdata.LocalExportAdapter;
 import com.grolinger.java.service.data.ApplicationDefinition;
 import com.grolinger.java.service.data.InterfaceDefinition;
 import com.grolinger.java.service.data.ServiceDefinition;
@@ -20,27 +20,24 @@ import java.io.Writer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.List;
 
 import static com.grolinger.java.controller.templatemodel.Constants.PATH_SEPARATOR;
 import static com.grolinger.java.controller.templatemodel.Constants.SLASH;
 
 @Slf4j
 @Service
-public class FileServiceImpl implements FileService {
+public class LocalExportAdapterImpl implements LocalExportAdapter {
     public static final String FILE_TYPE_IUML = ".iuml";
-    public static final String STATIC_INCLUDES = "static_includes";
+
     private static final String GLOBAL_FILE_EXPORT_PATH = System.getProperty("user.dir") + SLASH.getValue() + "target" + SLASH.getValue();
     private static final String COMMON_PATH = "common" + SLASH.getValue();
-    private static final String COMMON_FILE = "common" + SLASH.getValue() + "common.iuml";
     private static final String COMPONENT_DEFINITION_FILE = "component_definition.iuml";
     private static final String PARTICIPANT_DEFINITION_FILE = "participant_definition.iuml";
     private static final String EXAMPLE_FILE_SUFFIX = "_example.puml";
     private final SpringTemplateEngine templateEngine;
 
     @Autowired
-    private FileServiceImpl(SpringTemplateEngine templateEngine) {
+    private LocalExportAdapterImpl(SpringTemplateEngine templateEngine) {
         log.info("Configure base folder to read yaml files: {}", GLOBAL_FILE_EXPORT_PATH);
         this.templateEngine = templateEngine;
     }
@@ -60,55 +57,6 @@ public class FileServiceImpl implements FileService {
         path = GLOBAL_FILE_EXPORT_PATH + basePath + applicationPart + servicePart;
         Files.createDirectories(Paths.get(path));
         return path;
-    }
-
-    @Override
-    public void writeDefaultCommonFile(final String basePath, final DiagramType diagramType) throws IOException {
-        List<String> filesToExport;
-        //basePath already contains the diagram type
-        final String pathToExport = GLOBAL_FILE_EXPORT_PATH + basePath;
-        Files.createDirectories(Paths.get(pathToExport + COMMON_PATH));
-        log.info("Exporting base path {}", pathToExport);
-        if (DiagramType.COMPONENT_V1_2019_6_DIAGRAM_BASE.equals(diagramType) ||
-                DiagramType.COMPONENT_V1_2020_7_DIAGRAM_BASE.equals(diagramType)) {
-            filesToExport = Arrays.asList(
-                    COMMON_FILE,
-                    COMMON_PATH + "domain_legend.iuml",
-                    COMMON_PATH + "devicons" + SLASH.getValue() + "springBoot.puml",
-                    COMMON_PATH + "devicons" + SLASH.getValue() + "salesforce.puml",
-                    COMMON_PATH + "devicons" + SLASH.getValue() + "sap.puml",
-                    COMMON_PATH + "devicons" + SLASH.getValue() + "solr.puml",
-                    COMMON_PATH + "devicons" + SLASH.getValue() + "tibco.puml",
-                    "Readme.puml");
-            log.info("Write some common files: {}", filesToExport);
-            // FIXME: better solution than hardcoding here
-            Files.createDirectories(Paths.get(pathToExport + COMMON_PATH + "devicons"));
-        } else {
-            filesToExport = Arrays.asList(
-                    COMMON_FILE,
-                    COMMON_PATH + "bipro_messageIds.iuml",
-                    COMMON_PATH + "http_statuscodes.iuml");
-        }
-
-        for (String currentFileName : filesToExport) {
-            StaticFileService.copyFile(STATIC_INCLUDES + SLASH.getValue(),
-                    diagramType.getBasePath() + currentFileName);
-        }
-    }
-
-    @Override
-    public void writeDefaultSkinFiles() throws IOException {
-        final String subdir = "skin" + SLASH.getValue();
-        List<String> filesToExport = Arrays.asList(
-                subdir + "ci_company_colors.iuml",
-                subdir + "default.skin",
-                subdir + "darcula.skin");
-        log.info("Write some skin files: {}", filesToExport);
-        Files.createDirectories(Paths.get(GLOBAL_FILE_EXPORT_PATH + subdir));
-        //skin
-        for (String currentFileName : filesToExport) {
-            StaticFileService.copyFile(STATIC_INCLUDES + SLASH.getValue(), currentFileName);
-        }
     }
 
 
@@ -150,12 +98,12 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public String createDirectory(final String basePath, String path, final String applicationName) {
-            path = GLOBAL_FILE_EXPORT_PATH + basePath + StringUtils.capitalize(applicationName) + SLASH.getValue();
-            try {
-                Files.createDirectories(Paths.get(path));
-            } catch (IOException ioe) {
-                log.error("Could not create directory {}{} for application {}", basePath, path, applicationName);
-            }
+        path = GLOBAL_FILE_EXPORT_PATH + basePath + StringUtils.capitalize(applicationName) + SLASH.getValue();
+        try {
+            Files.createDirectories(Paths.get(path));
+        } catch (IOException ioe) {
+            log.error("Could not create directory {}{} for application {}", basePath, path, applicationName);
+        }
 
         return path;
     }